re-factor CI workflows (#42)
* re-factor CI workflows

* configure ci for master, develop branches
fix leveldb/util/env_win.cc and fs.cpp errors
enable qt, and tests

* remove altintegration.mk from depends

* fix Dockerfile

* fix Dockerfile

* fix Dockerfile

* disable alt-integration-cpp tests, set build type to release for MinGW build

* display src/test-suite.log for failed tests

* display src/test-suite.log for failed tests fix

* add alt-integration-cpp to depends, pull version from packages.mk

* fix echo statement

* fix echo statement

* update Dockerfile

* update Dockerfile

* fix permissions for install

* adjust CI for nightly full builds, add gui for master branch

* adjust CI builds

* remove tests for alt-integration-cpp

* fix Dockerfile, add verbose logs to cmake

* specify mingw32 compiler for cmake

* add coverage, adjust where we set env vars

* Use 'depends' for CI builds

* Fix

* Fix

* Fix

* Install deps

* Fix g++->gcc

* Revert "Fix g++->gcc"

This reverts commit 90a6ddb83b3e4b09f34426f7decb6663941d6aaf.

* Revert "Install deps"

This reverts commit b079d7420c1eb2f42dec89d89a01ed4340ff402f.

* Revert "Fix"

This reverts commit 30c6a2dbbade424f0ce863be2214d1c2c6579b2a.

* Revert "Fix"

This reverts commit 078e8a05fc76d3967c0976e9bb6dc219f2284dd2.

* Revert "Fix"

This reverts commit f4c5f8cea8ec1df41d17e2f7e7de76d604d1cbbc.

* Revert "Use 'depends' for CI builds"

This reverts commit 557382a7d5d2d99593a38b255c4a25459789a635.

* add cache for CI actions

* add cache for nightly CI builds

* remove cache from nightly CI builds (unsupported)

* test fix for ccache permissions

* use sudo for make for access to ccache

* use sudo for make install

* add timeout for make check, remove make check from mingw32 build (unsupported)

+++ src/fs.cpp
@@ -123,7 +123,7 @@ std::string get_filesystem_error_message(const fs::filesystem_error& e)
 
 static std::string openmodeToStr(std::ios_base::openmode mode)
 {
-    switch (mode & ~std::ios_base::ate) {
+    switch ((int)(mode & ~std::ios_base::ate)) {
     case std::ios_base::out:
     case std::ios_base::out | std::ios_base::trunc:
         return "w";
